requirements.txt
asgiref                  = 3.7.2
Django                   = 5.0.3
django-cors-headers      = 4.3.1
djangorestframework      = 3.14.0
mysqlclient              = 2.2.4
pillow                   = 10.2.0
pip                      = 24.0
pytz                     = 2024.1
rest-framework-simplejwt = 0.0.2
sqlparse                 = 0.4.4
tzdata                   = 2024.1